+++ patterns/dswp-information-contact-socials.php
@@ -1,0 +1,220 @@
+<?php
+/**
+ * Title: DSWP Information Contact Socials
+ * Slug: design-system-wordpress-theme/dswp-information-contact-socials
+ * Categories: contact
+ *
+ * @package Design-System-WordPress-Theme
+ */
+?>
+<!-- wp:group {"style":{"spacing":{"padding":{"right":"0","left":"0"}}},"layout":{"type":"constrained"}} -->
+<div class="wp-block-group" style="padding-right:0;padding-left:0"><!-- wp:group {"style":{"spacing":{"padding":{"right":"0","left":"0"}},"border":{"radius":"0px","color":"#d8d8d8","width":"1px"}},"className":"dswp-information-contact-socials-cards","layout":{"type":"constrained"}} -->
+    <div class="wp-block-group dswp-information-contact-socials-cards has-border-color" style="border-color:#d8d8d8;border-width:1px;border-radius:0px;padding-right:0;padding-left:0"><!-- wp:group {"style":{"color":{"background":"#ddedf2"},"spacing":{"padding":{"top":"var:preset|spacing|30","bottom":"var:preset|spacing|30"}}},"layout":{"type":"flex","orientation":"vertical"}} -->
+        <div class="wp-block-group has-background" style="background-color:#ddedf2;padding-top:var(--wp--preset--spacing--30);padding-bottom:var(--wp--preset--spacing--30)"><!-- wp:heading {"level":4,"style":{"spacing":{"padding":{"right":"var:preset|spacing|70","left":"var:preset|spacing|70"}}}} -->
+            <h4 class="wp-block-heading" style="padding-right:var(--wp--preset--spacing--70);padding-left:var(--wp--preset--spacing--70)">Neque porro quisquam</h4>
+            <!-- /wp:heading -->
+        </div>
+        <!-- /wp:group -->
+
+        <!-- wp:group {"style":{"spacing":{"padding":{"right":"var:preset|spacing|70","left":"var:preset|spacing|70"}}},"layout":{"type":"constrained"}} -->
+        <div class="wp-block-group" style="padding-right:var(--wp--preset--spacing--70);padding-left:var(--wp--preset--spacing--70)"><!-- wp:paragraph {"style":{"spacing":{"margin":{"top":"var:preset|spacing|50","bottom":"0"}}}} -->
+            <p style="margin-top:var(--wp--preset--spacing--50);margin-bottom:0">Lorem ipsum dolor sit amet, consectetur adipiscing elit. In quam nulla, blandit sed molestie sed, mattis rutrum sem. Praesent hendrerit arcu sed nibh blandit, quis fermentum odio posuere. Phasellus et varius eros. Integer eget erat vitae arcu tincidunt convallis aliquet id felis.</p>
+            <!-- /wp:paragraph -->
+
+            <!-- wp:group {"templateLock":false,"lock":{"move":false,"remove":false},"style":{"spacing":{"padding":{"top":"var:preset|spacing|50","bottom":"var:preset|spacing|50"}}},"className":"dswp-icon-with-excerpt","layout":{"type":"constrained"},"metadata":{"categories":["text"],"patternName":"design-system-wordpress-theme/dswp-horizontal-card","name":"DSWP Information Contact Socials"}} -->
+            <div class="wp-block-group dswp-icon-with-excerpt" style="padding-top:var(--wp--preset--spacing--50);padding-bottom:var(--wp--preset--spacing--50)"><!-- wp:columns {"style":{"spacing":{"margin":{"top":"0","bottom":"0"}}}} -->
+                <div class="wp-block-columns" style="margin-top:0;margin-bottom:0"><!-- wp:column {"style":{"spacing":{"padding":{"top":"0","bottom":"0"}}},"className":"dswp-information-contact-socials-card"} -->
+                    <div class="wp-block-column dswp-information-contact-socials-card" style="padding-top:0;padding-bottom:0"><!-- wp:group {"layout":{"type":"flex","orientation":"vertical"}} -->
+                        <div class="wp-block-group"><!-- wp:group {"style":{"spacing":{"padding":{"top":"0","bottom":"0"}}},"layout":{"type":"flex","flexWrap":"nowrap"}} -->
+                            <div class="wp-block-group" style="padding-top:0;padding-bottom:0"><!-- wp:image {"id":15470,"className":"size-full is-resized is-style-rounded"} -->
+                                <figure class="wp-block-image size-full is-resized is-style-rounded"><img src="<?php echo esc_url( get_template_directory_uri() . '/assets/images/square-256.png' ); ?>" alt=""/></figure>
+                                <!-- /wp:image -->
+
+                                <!-- wp:heading {"level":6,"style":{"spacing":{"margin":{"top":"0","bottom":"0"}}}} -->
+                                <h6 class="wp-block-heading" style="margin-top:0;margin-bottom:0">Neque porro quisquam</h6>
+                                <!-- /wp:heading -->
+                            </div>
+                            <!-- /wp:group -->
+
+                            <!-- wp:group {"style":{"spacing":{"padding":{"right":"0","left":"var:preset|spacing|80","top":"0"},"margin":{"top":"0","bottom":"0"}},"layout":{"selfStretch":"fit","flexSize":null}},"className":"dswp-information-contact-socials-card-content","layout":{"type":"flex","orientation":"vertical","verticalAlignment":"top"}} -->
+                            <div class="wp-block-group dswp-information-contact-socials-card-content" style="margin-top:0;margin-bottom:0;padding-top:0;padding-right:0;padding-left:var(--wp--preset--spacing--80)"><!-- wp:paragraph {"align":"left","style":{"layout":{"selfStretch":"fit","flexSize":null}}} -->
+                                <p class="has-text-align-left">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi bibendum rhoncus aliquam. Morbi at libero pellentesque, fermentum erat at, tincidunt nunc. Proin </p>
+                                <!-- /wp:paragraph -->
+
+                                <!-- wp:paragraph -->
+                                <p><a href="#test" data-type="internal" data-id="#test">Hyperlink</a></p>
+                                <!-- /wp:paragraph -->
+                            </div>
+                            <!-- /wp:group -->
+                        </div>
+                        <!-- /wp:group -->
+                    </div>
+                    <!-- /wp:column -->
+
+                    <!-- wp:column {"style":{"spacing":{"padding":{"top":"0","bottom":"0"}}},"className":"dswp-information-contact-socials-card"} -->
+                    <div class="wp-block-column dswp-information-contact-socials-card" style="padding-top:0;padding-bottom:0"><!-- wp:group {"layout":{"type":"flex","orientation":"vertical"}} -->
+                        <div class="wp-block-group"><!-- wp:group {"style":{"spacing":{"padding":{"top":"0","bottom":"0"}}},"layout":{"type":"flex","flexWrap":"nowrap"}} -->
+                            <div class="wp-block-group" style="padding-top:0;padding-bottom:0"><!-- wp:image {"id":15470,"className":"size-full is-resized is-style-rounded"} -->
+                                <figure class="wp-block-image size-full is-resized is-style-rounded"><img src="<?php echo esc_url( get_template_directory_uri() . '/assets/images/square-256.png' ); ?>" alt=""/></figure>
+                                <!-- /wp:image -->
+
+                                <!-- wp:heading {"level":6,"style":{"spacing":{"margin":{"top":"0","bottom":"0"}}}} -->
+                                <h6 class="wp-block-heading" style="margin-top:0;margin-bottom:0">Neque porro quisquam</h6>
+                                <!-- /wp:heading -->
+                            </div>
+                            <!-- /wp:group -->
+
+                            <!-- wp:group {"style":{"spacing":{"padding":{"right":"0","left":"var:preset|spacing|80","top":"0"},"margin":{"top":"0","bottom":"0"}},"layout":{"selfStretch":"fit","flexSize":null}},"className":"dswp-information-contact-socials-card-content","layout":{"type":"flex","orientation":"vertical","verticalAlignment":"top"}} -->
+                            <div class="wp-block-group dswp-information-contact-socials-card-content" style="margin-top:0;margin-bottom:0;padding-top:0;padding-right:0;padding-left:var(--wp--preset--spacing--80)"><!-- wp:paragraph {"align":"left","style":{"layout":{"selfStretch":"fit","flexSize":null}}} -->
+                                <p class="has-text-align-left">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi bibendum rhoncus aliquam. Morbi at libero pellentesque, fermentum erat at, tincidunt nunc. Proin </p>
+                                <!-- /wp:paragraph -->
+
+                                <!-- wp:paragraph -->
+                                <p><a href="#test" data-type="internal" data-id="#test">Hyperlink</a></p>
+                                <!-- /wp:paragraph -->
+                            </div>
+                            <!-- /wp:group -->
+                        </div>
+                        <!-- /wp:group -->
+                    </div>
+                    <!-- /wp:column -->
+
+                    <!-- wp:column {"style":{"spacing":{"padding":{"top":"0","bottom":"0"}}},"className":"dswp-information-contact-socials-card"} -->
+                    <div class="wp-block-column dswp-information-contact-socials-card" style="padding-top:0;padding-bottom:0"><!-- wp:group {"layout":{"type":"flex","orientation":"vertical"}} -->
+                        <div class="wp-block-group"><!-- wp:group {"style":{"spacing":{"padding":{"top":"0","bottom":"0"}}},"layout":{"type":"flex","flexWrap":"nowrap"}} -->
+                            <div class="wp-block-group" style="padding-top:0;padding-bottom:0"><!-- wp:image {"id":15470,"className":"size-full is-resized is-style-rounded"} -->
+                                <figure class="wp-block-image size-full is-resized is-style-rounded"><img src="<?php echo esc_url( get_template_directory_uri() . '/assets/images/square-256.png' ); ?>" alt=""/></figure>
+                                <!-- /wp:image -->
+
+                                <!-- wp:heading {"level":6,"style":{"spacing":{"margin":{"top":"0","bottom":"0"}}}} -->
+                                <h6 class="wp-block-heading" style="margin-top:0;margin-bottom:0">Neque porro quisquam</h6>
+                                <!-- /wp:heading -->
+                            </div>
+                            <!-- /wp:group -->
+
+                            <!-- wp:group {"style":{"spacing":{"padding":{"right":"0","left":"var:preset|spacing|80","top":"0"},"margin":{"top":"0","bottom":"0"}},"layout":{"selfStretch":"fit","flexSize":null}},"className":"dswp-information-contact-socials-card-content","layout":{"type":"flex","orientation":"vertical","verticalAlignment":"top"}} -->
+                            <div class="wp-block-group dswp-information-contact-socials-card-content" style="margin-top:0;margin-bottom:0;padding-top:0;padding-right:0;padding-left:var(--wp--preset--spacing--80)"><!-- wp:paragraph {"align":"left","style":{"layout":{"selfStretch":"fit","flexSize":null}}} -->
+                                <p class="has-text-align-left">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi bibendum rhoncus aliquam. Morbi at libero pellentesque, fermentum erat at, tincidunt nunc. Proin </p>
+                                <!-- /wp:paragraph -->
+
+                                <!-- wp:paragraph -->
+                                <p><a href="#test" data-type="internal" data-id="#test">Hyperlink</a></p>
+                                <!-- /wp:paragraph -->
+                            </div>
+                            <!-- /wp:group -->
+                        </div>
+                        <!-- /wp:group -->
+                    </div>
+                    <!-- /wp:column -->
+                </div>
+                <!-- /wp:columns -->
+
+                <!-- wp:columns {"style":{"spacing":{"margin":{"top":"0","bottom":"0"}}}} -->
+                <div class="wp-block-columns" style="margin-top:0;margin-bottom:0"><!-- wp:column {"style":{"spacing":{"padding":{"top":"0","bottom":"0"}}},"className":"dswp-information-contact-socials-card"} -->
+                    <div class="wp-block-column dswp-information-contact-socials-card" style="padding-top:0;padding-bottom:0"><!-- wp:group {"layout":{"type":"flex","orientation":"vertical"}} -->
+                        <div class="wp-block-group"><!-- wp:group {"style":{"spacing":{"padding":{"top":"0","bottom":"0"}}},"layout":{"type":"flex","flexWrap":"nowrap"}} -->
+                            <div class="wp-block-group" style="padding-top:0;padding-bottom:0"><!-- wp:image {"id":15470,"className":"size-full is-resized is-style-rounded"} -->
+                                <figure class="wp-block-image size-full is-resized is-style-rounded"><img src="<?php echo esc_url( get_template_directory_uri() . '/assets/images/square-256.png' ); ?>" alt=""/></figure>
+                                <!-- /wp:image -->
+
+                                <!-- wp:heading {"level":6,"style":{"spacing":{"margin":{"top":"0","bottom":"0"}}}} -->
+                                <h6 class="wp-block-heading" style="margin-top:0;margin-bottom:0">Neque porro quisquam</h6>
+                                <!-- /wp:heading -->
+                            </div>
+                            <!-- /wp:group -->
+
+                            <!-- wp:group {"style":{"spacing":{"padding":{"right":"0","left":"var:preset|spacing|80","top":"0"},"margin":{"top":"0","bottom":"0"}},"layout":{"selfStretch":"fit","flexSize":null}},"className":"dswp-information-contact-socials-card-content","layout":{"type":"flex","orientation":"vertical","verticalAlignment":"top"}} -->
+                            <div class="wp-block-group dswp-information-contact-socials-card-content" style="margin-top:0;margin-bottom:0;padding-top:0;padding-right:0;padding-left:var(--wp--preset--spacing--80)"><!-- wp:paragraph {"align":"left","style":{"layout":{"selfStretch":"fit","flexSize":null}}} -->
+                                <p class="has-text-align-left">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi bibendum rhoncus aliquam. Morbi at libero pellentesque, fermentum erat at, tincidunt nunc. Proin </p>
+                                <!-- /wp:paragraph -->
+
+                                <!-- wp:paragraph -->
+                                <p><a href="#test" data-type="internal" data-id="#test">Hyperlink</a></p>
+                                <!-- /wp:paragraph -->
+                            </div>
+                            <!-- /wp:group -->
+                        </div>
+                        <!-- /wp:group -->
+                    </div>
+                    <!-- /wp:column -->
+
+                    <!-- wp:column {"style":{"spacing":{"padding":{"top":"0","bottom":"0"}}},"className":"dswp-information-contact-socials-card"} -->
+                    <div class="wp-block-column dswp-information-contact-socials-card" style="padding-top:0;padding-bottom:0"><!-- wp:group {"layout":{"type":"flex","orientation":"vertical"}} -->
+                        <div class="wp-block-group"><!-- wp:group {"style":{"spacing":{"padding":{"top":"0","bottom":"0"}}},"layout":{"type":"flex","flexWrap":"nowrap"}} -->
+                            <div class="wp-block-group" style="padding-top:0;padding-bottom:0"><!-- wp:image {"id":15470,"className":"size-full is-resized is-style-rounded"} -->
+                                <figure class="wp-block-image size-full is-resized is-style-rounded"><img src="<?php echo esc_url( get_template_directory_uri() . '/assets/images/square-256.png' ); ?>" alt=""/></figure>
+                                <!-- /wp:image -->
+
+                                <!-- wp:heading {"level":6,"style":{"spacing":{"margin":{"top":"0","bottom":"0"}}}} -->
+                                <h6 class="wp-block-heading" style="margin-top:0;margin-bottom:0">Neque porro quisquam</h6>
+                                <!-- /wp:heading -->
+                            </div>
+                            <!-- /wp:group -->
+
+                            <!-- wp:group {"style":{"spacing":{"padding":{"right":"0","left":"var:preset|spacing|80","top":"0"},"margin":{"top":"0","bottom":"0"}},"layout":{"selfStretch":"fit","flexSize":null}},"className":"dswp-information-contact-socials-card-content","layout":{"type":"flex","orientation":"vertical","verticalAlignment":"top"}} -->
+                            <div class="wp-block-group dswp-information-contact-socials-card-content" style="margin-top:0;margin-bottom:0;padding-top:0;padding-right:0;padding-left:var(--wp--preset--spacing--80)"><!-- wp:paragraph {"align":"left","style":{"layout":{"selfStretch":"fit","flexSize":null}}} -->
+                                <p class="has-text-align-left">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi bibendum rhoncus aliquam. Morbi at libero pellentesque, fermentum erat at, tincidunt nunc. Proin </p>
+                                <!-- /wp:paragraph -->
+
+                                <!-- wp:paragraph -->
+                                <p><a href="#test" data-type="internal" data-id="#test">Hyperlink</a></p>
+                                <!-- /wp:paragraph -->
+                            </div>
+                            <!-- /wp:group -->
+                        </div>
+                        <!-- /wp:group -->
+                    </div>
+                    <!-- /wp:column -->
+
+                    <!-- wp:column {"style":{"spacing":{"padding":{"top":"0","bottom":"0"}}},"className":"dswp-information-contact-socials-card"} -->
+                    <div class="wp-block-column dswp-information-contact-socials-card" style="padding-top:0;padding-bottom:0"><!-- wp:group {"layout":{"type":"flex","orientation":"vertical"}} -->
+                        <div class="wp-block-group"><!-- wp:group {"style":{"spacing":{"padding":{"top":"0","bottom":"0"}}},"layout":{"type":"flex","flexWrap":"nowrap"}} -->
+                            <div class="wp-block-group" style="padding-top:0;padding-bottom:0"><!-- wp:image {"id":15470,"className":"size-full is-resized is-style-rounded"} -->
+                                <figure class="wp-block-image size-full is-resized is-style-rounded"><img src="<?php echo esc_url( get_template_directory_uri() . '/assets/images/square-256.png' ); ?>" alt=""/></figure>
+                                <!-- /wp:image -->
+
+                                <!-- wp:heading {"level":6,"style":{"spacing":{"margin":{"top":"0","bottom":"0"}}}} -->
+                                <h6 class="wp-block-heading" style="margin-top:0;margin-bottom:0">Neque porro quisquam</h6>
+                                <!-- /wp:heading -->
+                            </div>
+                            <!-- /wp:group -->
+
+                            <!-- wp:group {"style":{"spacing":{"padding":{"right":"0","left":"var:preset|spacing|80","top":"0"},"margin":{"top":"0","bottom":"0"}},"layout":{"selfStretch":"fit","flexSize":null}},"className":"dswp-information-contact-socials-card-content","layout":{"type":"flex","orientation":"vertical","verticalAlignment":"top"}} -->
+                            <div class="wp-block-group dswp-information-contact-socials-card-content" style="margin-top:0;margin-bottom:0;padding-top:0;padding-right:0;padding-left:var(--wp--preset--spacing--80)"><!-- wp:paragraph {"align":"left","style":{"layout":{"selfStretch":"fit","flexSize":null}}} -->
+                                <p class="has-text-align-left">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi bibendum rhoncus aliquam. Morbi at libero pellentesque, fermentum erat at, tincidunt nunc. Proin </p>
+                                <!-- /wp:paragraph -->
+
+                                <!-- wp:paragraph -->
+                                <p><a href="#test" data-type="internal" data-id="#test">Hyperlink</a></p>
+                                <!-- /wp:paragraph -->
+                            </div>
+                            <!-- /wp:group -->
+                        </div>
+                        <!-- /wp:group -->
+                    </div>
+                    <!-- /wp:column -->
+                </div>
+                <!-- /wp:columns -->
+            </div>
+            <!-- /wp:group -->
+
+            <!-- wp:buttons {"layout":{"type":"flex","justifyContent":"center"}} -->
+            <div class="wp-block-buttons"><!-- wp:button {"className":"is-style-outline"} -->
+                <div class="wp-block-button is-style-outline"><a class="wp-block-button__link wp-element-button">Vestibulum sit amet</a></div>
+                <!-- /wp:button -->
+            </div>
+            <!-- /wp:buttons -->
+
+            <!-- wp:spacer {"height":"22px"} -->
+            <div style="height:22px" aria-hidden="true" class="wp-block-spacer"></div>
+            <!-- /wp:spacer -->
+        </div>
+        <!-- /wp:group -->
+    </div>
+    <!-- /wp:group -->
+
+    <!-- wp:spacer {"height":"17px"} -->
+    <div style="height:17px" aria-hidden="true" class="wp-block-spacer"></div>
+    <!-- /wp:spacer -->
+</div>
+<!-- /wp:group -->